+++ apollo-runtime/src/main/java/com/apollographql/android/RealApolloCall.java
@@ -4,6 +4,7 @@ import com.apollographql.android.api.graphql.Operation;
 import com.apollographql.android.api.graphql.Response;
 import com.apollographql.android.api.graphql.ResponseFieldMapper;
 import com.apollographql.android.api.graphql.ScalarType;
+import com.apollographql.android.cache.HttpCacheInterceptor;
 import com.squareup.moshi.JsonAdapter;
 import com.squareup.moshi.Moshi;
