+++ apollo-runtime/src/main/java/com/apollographql/android/cache/ResponseCacheStore.java
@@ -1,0 +1,13 @@
+package com.apollographql.android.cache;
+
+import java.io.IOException;
+
+import javax.annotation.Nonnull;
+
+public interface ResponseCacheStore {
+  ResponseCacheRecord cacheRecord(@Nonnull String cacheKey) throws IOException;
+
+  ResponseCacheRecordEditor cacheRecordEditor(@Nonnull String cacheKey) throws IOException;
+
+  void delete() throws IOException;
+}
